+++ my_project/tab_outdoor_comfort/app_outdoor_comfort.py
@@ -5,7 +5,7 @@ import dash_bootstrap_components as dbc
 from my_project.global_scheme import outdoor_dropdown_names, tight_margins, month_lst, container_col_center_one_of_three
 from dash.dependencies import Input, Output, State
 from my_project.template_graphs import heatmap_with_filter, thermal_stress_stacked_barchart
-from my_project.utils import title_with_tooltip, generate_chart_name
+from my_project.utils import title_with_tooltip, generate_chart_name, generate_units_degree, generate_units
 import numpy as np
 from app import app
 
@@ -174,10 +174,11 @@ def layout_outdoor_comfort():
     ],
 )
 def update_tab_utci_value(ts, var, global_local, time_filter, df, meta, si_ip, month, hour, invert_month, invert_hour):
-
+    custom_inputs = f"{var}"
+    units = generate_units_degree(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("utci_heatmap", meta),
-        figure=heatmap_with_filter(df, var, global_local, si_ip, time_filter, month, hour, invert_month, invert_hour),
+        config=generate_chart_name("heatmap", meta, custom_inputs, units),
+        figure=heatmap_with_filter(df, var, global_local, si_ip, time_filter, month, hour, invert_month, invert_hour)
     )
 
 
@@ -222,7 +223,7 @@ def update_tab_utci_category(ts, var, global_local, time_filter, df, meta, si_ip
         title="Thermal stress",
         titleside="top",
         tickmode="array",
-        tickvals=[4, 3, 2, 1, 0, -1, -2, -3, -4, -5],
+        tickvals = np.linspace(-4.75, 4.75, 10),
         ticktext=[
             "extreme heat stress",
             "very strong heat stress",
@@ -237,8 +238,10 @@ def update_tab_utci_category(ts, var, global_local, time_filter, df, meta, si_ip
         ],
         ticks="outside",
     )
+    custom_inputs = f"{var}"
+    units = generate_units(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("utci_heatmap_category", meta),
+        config=generate_chart_name("heatmap_category", meta, custom_inputs, units),
         figure=utci_stress_cat,
     )
 
@@ -254,7 +257,8 @@ def update_tab_utci_category(ts, var, global_local, time_filter, df, meta, si_ip
         State("outdoor-comfort-hour-slider", "value"),
         State("meta-store", "data"),
         State("invert-month-outdoor-comfort", "value"),
-        State("invert-hour-outdoor-comfort", "value")
+        State("invert-hour-outdoor-comfort", "value"),
+        State("si-ip-unit-store", "data")
     ],
 )
 def update_tab_utci_summary_chart(
@@ -265,7 +269,8 @@ def update_tab_utci_summary_chart(
         hour,
         meta,
         invert_month,
-        invert_hour
+        invert_hour,
+        si_ip
 ):
     utci_summary_chart = thermal_stress_stacked_barchart(df, var + "_categories", time_filter, month, hour, invert_month, invert_hour)
     utci_summary_chart = utci_summary_chart.update_layout(
@@ -276,7 +281,9 @@ def update_tab_utci_summary_chart(
     utci_summary_chart.update_xaxes(
         dict(tickmode="array", tickvals=np.arange(0, 12, 1), ticktext=month_lst)
     )
+    custom_inputs = f"{var}"
+    units = generate_units(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("utci_summary_chart", meta),
+        config=generate_chart_name("summary", meta, custom_inputs, units),
         figure=utci_summary_chart,
     )