+++ my_project/tab_sun/app_sun.py
@@ -1,10 +1,13 @@
 import numpy as np
+from copy import deepcopy
 from dash import dcc
 from dash import html
 import dash_bootstrap_components as dbc
 from my_project.global_scheme import (
     sun_cloud_tab_dropdown_names,
     sun_cloud_tab_explore_dropdown_names,
+    dropdown_names,
+    more_variables_dropdown,
     tight_margins,
     month_lst,
     mapping_dictionary,
@@ -18,9 +21,24 @@ from my_project.tab_sun.charts_sun import (
 )
 from my_project.template_graphs import heatmap, barchart, daily_profile
 from my_project.utils import code_timer
-from my_project.utils import title_with_tooltip, generate_chart_name
+from my_project.utils import title_with_tooltip, generate_chart_name, generate_units, generate_custom_inputs
 
 from app import app
+
+sc_dropdown_names = {
+    "None": "None",
+    "Frequency": "Frequency",
+}
+sc_dropdown_names.update(deepcopy(dropdown_names))
+sc_dropdown_names.update(deepcopy(sun_cloud_tab_dropdown_names))
+sc_dropdown_names.update(deepcopy(sun_cloud_tab_explore_dropdown_names))
+# Remove the keys from the dictionary
+sc_dropdown_names.pop("Vapor partial pressure", None)
+sc_dropdown_names.pop("Absolute humidity", None)
+sc_dropdown_names.pop("UTCI: Sun & Wind : categories", None)
+sc_dropdown_names.pop("UTCI: no Sun & Wind : categories", None)
+sc_dropdown_names.pop("UTCI: Sun & no Wind : categories", None)
+sc_dropdown_names.pop("UTCI: no Sun & no Wind : categories", None)
 
 
 def sun_path():
@@ -67,8 +85,8 @@ def sun_path():
                     dcc.Dropdown(
                         id="custom-sun-var-dropdown",
                         options=[
-                            {"label": i, "value": sun_cloud_tab_dropdown_names[i]}
-                            for i in sun_cloud_tab_dropdown_names
+                            {"label": i, "value": sc_dropdown_names[i]}
+                            for i in sc_dropdown_names
                         ],
                         value="None",
                         style={"width": "20rem"},
@@ -205,12 +223,12 @@ def monthly_and_cloud_chart(ts, df, meta, si_ip):
     cover.update_xaxes(
         dict(tickmode="array", tickvals=np.arange(0, 12, 1), ticktext=month_lst)
     )
-
+    units = generate_units(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("monthly_sun", meta),
+        config=generate_chart_name("Global_and_Diffuse_Horizontal_Solar_Radiation", meta, units),
         figure=monthly,
     ), dcc.Graph(
-        config=generate_chart_name("cloud_cover_sun", meta),
+        config=generate_chart_name("cloud_cover", meta, units),
         figure=cover,
     )
 
@@ -232,15 +250,16 @@ def monthly_and_cloud_chart(ts, df, meta, si_ip):
 @code_timer
 def sun_path_chart(ts, view, var, global_local, df, meta, si_ip):
     """Update the contents of tab four. Passing in the polar selection and the general info (df, meta)."""
-
+    custom_inputs = "" if var == "None" else f"{var}"
+    units = "" if var == "None" else generate_units(si_ip)
     if view == "polar":
         return dcc.Graph(
-            config=generate_chart_name("spherical_sun_path_sun", meta),
+            config=generate_chart_name("spherical_sunpath", meta, custom_inputs, units),
             figure=polar_graph(df, meta, global_local, var, si_ip),
         )
     else:
         return dcc.Graph(
-            config=generate_chart_name("cartesian_sun_path_sun", meta),
+            config=generate_chart_name("cartesian_sunpath", meta, custom_inputs, units),
             figure=custom_cartesian_solar(df, meta, global_local, var, si_ip),
         )
 
@@ -261,9 +280,10 @@ def sun_path_chart(ts, view, var, global_local, df, meta, si_ip):
 @code_timer
 def daily(ts, var, global_local, df, meta, si_ip):
     """Update the contents of tab four section two. Passing in the general info (df, meta)."""
-
+    custom_inputs = generate_custom_inputs(var)
+    units = generate_units(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("daily_sun", meta),
+        config=generate_chart_name("daily", meta, custom_inputs, units),
         figure=daily_profile(df, var, global_local, si_ip),
     )
 
@@ -283,8 +303,9 @@ def daily(ts, var, global_local, df, meta, si_ip):
 )
 @code_timer
 def update_heatmap(ts, var, global_local, df, meta, si_ip):
-
+    custom_inputs = generate_custom_inputs(var)
+    units = generate_units(si_ip)
     return dcc.Graph(
-        config=generate_chart_name("heatmap_sun", meta),
+        config=generate_chart_name("heatmap", meta, custom_inputs, units),
         figure=heatmap(df, var, global_local, si_ip),
     )